Serializers added View implemented
