+++ django_view_test/urls.py
@@ -17,6 +17,7 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path("api/v1/foods/", include("foods.urls")),
+    path("api/v1/", include("foods.urls"), name="foods"),
+    path("api-auth/", include("rest_framework.urls")),
     path("admin/", admin.site.urls),
 ]
