+++ foods/urls.py
@@ -1,5 +1,7 @@
 from django.urls import path
 
-from . import views
+from foods import views
 
-urlpatterns = [path("", views.foods, name="foods")]
+urlpatterns = [
+    path("foods/", views.FoodsList.as_view(), name="foods list"),
+]
